+++ dungen/game/__init__.py
@@ -1,0 +1,3 @@
+from .core import Game
+
+__all__ = ['Game']
