+++ fantasy.yaml
@@ -9,7 +9,7 @@ game_settings:
   character_panel_color: "bright_cyan"
   status_panel_color: "bright_cyan"
   map_panel_color: "cyan"
-  message_history_limit: 10
+  message_history_limit: 8
   recent_encounters_limit: 6
 
 player:
